Moves projected HostTelemetry from models to entities
